Moved `ExtractFieldFromJWT` to common `utils` package
